--- src/stories/ColorComponent.stories.tsx
@@ -1,15 +1,0 @@
-import { Story } from '@storybook/react';
-import ColorComponent, { ColorComponentProps } from './ColorComponent';
-
-export default {
-  title: 'Themes/Color',
-  component: ColorComponent,
-};
-
-const Template: Story<ColorComponentProps> = (props) => <ColorComponent {...props} />;
-
-export const Primary = Template.bind({});
-Primary.args = {
-  colorKey: 'background',
-  theme: 'light',
-};
