+++ src/themes/color.stories.tsx
@@ -1,9 +1,14 @@
 import type { HTMLProps } from 'react';
 import type { Color, ColorTheme } from '@themes/color';
+// types
+
+import { Story } from '@storybook/react';
+// React.js module
 
 import { colorMap } from '@themes/color';
+// styles
 
-export interface ColorComponentProps extends HTMLProps<HTMLDivElement> {
+interface ColorComponentProps extends HTMLProps<HTMLDivElement> {
   colorKey: keyof Color;
   theme: ColorTheme;
 }
@@ -22,4 +27,15 @@ const ColorComponent = ({ colorKey, theme, ...props }: ColorComponentProps) => {
   );
 };
 
-export default ColorComponent;
+export default {
+  title: 'Themes/Color',
+  component: ColorComponent,
+};
+
+const Template: Story<ColorComponentProps> = (props) => <ColorComponent {...props} />;
+
+export const Primary = Template.bind({});
+Primary.args = {
+  colorKey: 'background',
+  theme: 'light',
+};
